app/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bson==0.5.10
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.1
narwhals==2.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
passlib==1.7.4
plotly==6.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1
python-multipart==0.0.6
pytest-cov==4.1.0